etl/dags/job_scraping_dag.py
"""
TrendCatcher ETL Pipeline - Job Scraping DAG

Bu DAG her 2 saatte bir çalışarak:
1. LinkedIn, Indeed ve Kariyer.net'ten iş ilanlarını toplar
2. Verileri normalize eder  
3. NLP ile beceri çıkarımı yapar
4. TimescaleDB'ye kaydeder

ETL TEAM: Bu dosyayı kendi ihtiyaçlarınıza göre geliştirin!
"""

from datetime import datetime, timedelta
from airflow import DAG
from airflow.operators.python import PythonOperator
from airflow.providers.postgres.operators.postgres import PostgresOperator
from airflow.utils.dates import days_ago

# TODO: ETL Team - Kendi scraper modüllerinizi import edin
# from scripts.linkedin_scraper import scrape_linkedin_jobs
# from scripts.indeed_scraper import scrape_indeed_jobs
# from scripts.kariyer_scraper import scrape_kariyer_jobs
# from scripts.data_processor import normalize_job_data, extract_skills
# from scripts.database_utils import save_to_timescaledb

default_args = {
    'owner': 'trendcatcher-etl-team',
    'depends_on_past': False,
    'start_date': days_ago(1),
    'email_on_failure': True,  # TODO: Ekip email'lerini ekleyin
    'email_on_retry': False,
    'retries': 2,
    'retry_delay': timedelta(minutes=15),
    'max_active_runs': 1,
}

dag = DAG(
    'job_scraping_pipeline',
    default_args=default_args,
    description='İş ilanlarını toplayan ve işleyen ETL pipeline',
    schedule_interval=timedelta(hours=2),  # Her 2 saatte bir
    catchup=False,
    tags=['etl', 'scraping', 'jobs'],
)

def scrape_linkedin_task(**context):
    """
    LinkedIn Scraping Task
    
    TODO: ETL Team - Bu fonksiyonu implement edin:
    
    PSEUDO CODE:
    1. Search terms listesi oluştur (Data Scientist, Software Engineer, etc.)
    2. Target cities listesi (Istanbul, Ankara, Izmir, etc.)
    3. Her search term + city kombinasyonu için:
       - LinkedIn'e request gönder
       - Rate limiting uygula (1-3 saniye arası)
       - HTML parse et (BeautifulSoup/Playwright)
       - Job details çıkar:
         * title, company, location
         * salary (eğer varsa)
         * job description
         * posting date
         * job_url
    4. Duplicate check yap
    5. XCom ile sonraki task'a aktar
    
    RETURN: Scrape edilen job sayısı
    """
    # TODO: Implement LinkedIn scraping logic
    pass


def scrape_indeed_task(**context):
    """
    Indeed Scraping Task
    
    TODO: ETL Team - Bu fonksiyonu implement edin:
    
    PSEUDO CODE:
    1. Indeed API key kullan (eğer varsa) veya web scrape
    2. Türkçe arama terimleri (veri bilimci, yazılım geliştirici)
    3. Her search için:
       - API call veya web request
       - Response parse et
       - Job data extract et
    4. Rate limiting dikkat et
    5. XCom push
    """
    # TODO: Implement Indeed scraping logic
    pass


def scrape_kariyer_task(**context):
    """
    Kariyer.net Scraping Task
    
    TODO: ETL Team - Bu fonksiyonu implement edin:
    
    PSEUDO CODE:
    1. Kariyer.net için özel scraping stratejisi
    2. Turkish job market focus
    3. Salary information dikkatli extract et
    4. Company details al
    5. XCom push
    """
    # TODO: Implement Kariyer.net scraping logic
    pass


def normalize_data_task(**context):
    """
    Data Normalization Task
    
    TODO: ETL Team - Data normalization logic:
    
    PSEUDO CODE:
    1. XCom'dan 3 kaynaktan data çek
    2. Her kaynak için common schema'ya convert et:
       - Standardize field names
       - Clean text data
       - Parse salary strings to numbers
       - Normalize location names
       - Extract posting dates
    3. Duplicate detection (title + company + date)
    4. Data validation
    5. XCom push normalized data
    """
    # TODO: Implement data normalization
    pass


def extract_skills_task(**context):
    """
    NLP Skill Extraction Task
    
    TODO: ML/NLP Team ile koordine edin:
    
    PSEUDO CODE:
    1. Normalized job descriptions al
    2. Her description için:
       - Text preprocessing (clean, tokenize)
       - NER model ile skill extraction
       - Custom skill keywords match
       - Technology stack detection
    3. Skill confidence scores
    4. Skill standardization (Python = python = PYTHON)
    5. XCom push jobs with skills
    """
    # TODO: Implement NLP skill extraction
    pass


def save_to_database_task(**context):
    """
    Database Save Task
    
    TODO: ETL Team - Database operations:
    
    PSEUDO CODE:
    1. Jobs with skills data al
    2. TimescaleDB connection aç
    3. Her job için:
       - Duplicate check (son 24 saat)
       - Insert job_postings table
       - Insert skills relationship table
       - Update company table
    4. Batch insert for performance
    5. Error handling ve logging
    6. Return saved count
    """
    # TODO: Implement database save logic
    pass


def generate_metrics_task(**context):
    """
    Metrics Generation Task
    
    TODO: DevOps Team - Monitoring metrics:
    
    PSEUDO CODE:
    1. Pipeline success metrics
    2. Job count per source
    3. Skill distribution
    4. Error rates
    5. Processing time
    6. Prometheus metrics push
    """
    # TODO: Implement metrics generation
    pass


# Task definitions
scrape_linkedin = PythonOperator(
    task_id='scrape_linkedin',
    python_callable=scrape_linkedin_task,
    dag=dag,
)

scrape_indeed = PythonOperator(
    task_id='scrape_indeed', 
    python_callable=scrape_indeed_task,
    dag=dag,
)

scrape_kariyer = PythonOperator(
    task_id='scrape_kariyer',
    python_callable=scrape_kariyer_task,
    dag=dag,
)

normalize_data = PythonOperator(
    task_id='normalize_data',
    python_callable=normalize_data_task,
    dag=dag,
)

extract_skills = PythonOperator(
    task_id='extract_skills',
    python_callable=extract_skills_task,
    dag=dag,
)

save_to_database = PythonOperator(
    task_id='save_to_database',
    python_callable=save_to_database_task,
    dag=dag,
)

generate_metrics = PythonOperator(
    task_id='generate_metrics',
    python_callable=generate_metrics_task,
    dag=dag,
)

# TODO: ETL Team - Database schema'yı ihtiyaçlarınıza göre güncelleyin
create_tables = PostgresOperator(
    task_id='create_tables',
    postgres_conn_id='timescaledb_default',
    sql="""
    -- TODO: ETL Team - Bu SQL'i geliştirin
    
    -- Ana job postings tablosu (TimescaleDB hypertable)
    CREATE TABLE IF NOT EXISTS job_postings (
        id SERIAL PRIMARY KEY,
        scraped_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
        job_title VARCHAR(500) NOT NULL,
        company VARCHAR(300) NOT NULL,
        -- TODO: Diğer field'ları ekleyin
    );
    
    -- Hypertable'a çevir (time-series için)
    SELECT create_hypertable('job_postings', 'scraped_at', if_not_exists => TRUE);
    
    -- TODO: İhtiyaçlarınıza göre indexler ekleyin
    """,
    dag=dag,
)

# Task dependencies - ETL Team bu flow'u optimize edebilir
create_tables >> [scrape_linkedin, scrape_indeed, scrape_kariyer]
[scrape_linkedin, scrape_indeed, scrape_kariyer] >> normalize_data
normalize_data >> extract_skills  
extract_skills >> save_to_database
save_to_database >> generate_metrics 